.history/backend/requirements_20260113223819.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
python-docx==1.1.2
